requirements.txt
streamlit
pandas
numpy
matplotlib
python-pptx
openpyxl
Pillow